% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/drp.R
\name{drp}
\alias{drp}
\title{drp}
\usage{
drp(GSmodel10, NSmodel, NSSmodel, avspotratesa)
}
\arguments{
\item{GSmodel10}{a GaussianSmooth object}

\item{NSmodel}{a NelsonSiegel object}

\item{NSSmodel}{a NelsonSiegelSvensson object}

\item{avspotratesa}{average swap rates (semi-annual basis)}
}
\value{
list consisting of Average Swap rate (semi-annual basis), Annualized Swap average, 10 year final cost of debt estimate","10 year DRP"
}
\description{
Determine Debt Risk Premium
}

