% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/drpadjust.R
\name{drpadjust}
\alias{drpadjust}
\title{drpadjust}
\usage{
drpadjust(GSadjustmodel10, NSadjustmodel, NSSadjustmodel, avspotratesa)
}
\arguments{
\item{avspotratesa}{average swap rates (semi-annual basis)}

\item{data}{Data file consisting of at least 4 columns: Maturity, Amount.issued, Yield, and Credit.rating}

\item{GSmodel10}{a GaussianSmooth object}

\item{NSmodel}{a NelsonSiegel object}

\item{NSSmodel}{a NelsonSiegelSvensson object}
}
\value{
list consisting of Average Swap rate (semi-annual basis), Annualized Swap average, 10 year final cost of debt estimate","10 year DRP"
}
\description{
Determine Debt Risk Premium adjusted for Credit ratings
}

