% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/drpplot.R
\name{drpplot}
\alias{drpplot}
\title{drplot}
\usage{
drpplot(GSmodel10, NSmodel, NSSmodel)
}
\arguments{
\item{NSmodel}{}

\item{NSSmodel}{}

\item{GS10}{}
}
\value{
scatter plot of Yields vs Maturities with Gaussian Smooth, Nelson-Siegel, and Nelson-Siegel-Svensson curves
}
\description{
Compare Gaussian Smoothing, Nelson-Siegel, and Nelson-Siegel-Svensson fits
}

