% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitnelsonsiegeladjust.R
\name{fitnelsonsiegeladjust}
\alias{fitnelsonsiegeladjust}
\title{Fit the Nelson-Siegel model adjusting for Credit ratings}
\usage{
fitnelsonsiegeladjust(data, lambdastartval = lambdastart(2.5),
  epsilon = 1e-14)
}
\arguments{
\item{data}{data file with at least columns named Yield and Maturity and Credit rating}

\item{lambdastartval}{starting value for lambda in Nelson-Siegel model}

\item{epsilon}{a small value to prevent -Inf when taking log of zero}
}
\value{
A NelsonSiegelAdjust object
}
\description{
Fit the Nelson-Siegel model adjusting for Credit ratings
}

