% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitnelsonsiegelsvensson.R
\name{fitnelsonsiegelsvensson}
\alias{fitnelsonsiegelsvensson}
\title{Fit the Nelson-Siegel-Svensson model}
\usage{
fitnelsonsiegelsvensson(data, lambda1start = 1/lambdastart(2.5),
  lambda2start = 1/lambdastart(5), epsilon = 1e-14)
}
\arguments{
\item{data}{data file with at least columns named Yield and Maturity}

\item{lambda1start}{starting value for lambda1}

\item{lambda2start}{starting value for lambda2; lambda2 > lambda1}

\item{epsilon}{a small value to prevent -Inf when taking log of zero}
}
\value{
A NelsonSiegelSvensson object
}
\description{
Fit the Nelson-Siegel-Svensson model
}

