% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitnelsonsiegelsvenssonadjust.R
\name{fitnelsonsiegelsvenssonadjust}
\alias{fitnelsonsiegelsvenssonadjust}
\title{Fit the Nelson-Siegel-Svensson model adjusting for Credit ratings}
\usage{
fitnelsonsiegelsvenssonadjust(data, lambda1start = 1/lambdastart(2.5),
  lambda2start = 1/lambdastart(5), epsilon = 1e-14)
}
\arguments{
\item{data}{data file with at least columns named Yield and Maturity}

\item{epsilon}{a small value to prevent -Inf when taking log of zero}

\item{loglambda1start}{starting value for log(lambda1) in Nelson-Siegel-Svensson model}

\item{loglambda2start}{starting value for log(lambda2) in Nelson-Siegel-Svensson model}
}
\value{
A NelsonSiegelSvenssonAdjust object
}
\description{
Fit the Nelson-Siegel-Svensson model adjusting for Credit ratings
}

