% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gkest1.R
\name{gkest1}
\alias{gkest1}
\title{gkest}
\usage{
gkest1(sigma = 1.5, data = Jan15)
}
\arguments{
\item{sigma}{Smoothing Constant}

\item{data}{Data file consisting of at least 3 columns: Maturity, Amount.issued, and Yield}
}
\value{
list consisting of a matrix of Target Tenor, sigma, Target Tenor Yield, Effective Term to Maturity, and Annual Yield; Interpolated 7 year yield (Semi-annual basis); Extrapolated 10 year yield (Semi-annual basis); Interpolated 7 year yield (Annual basis); and Extrapolated 10 year yield (Annual basis)
}
\description{
Interpolated and Extrapolated 7 and 10 year yields
}

