% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gkest1adjust.R
\name{gkest1adjust}
\alias{gkest1adjust}
\title{gkest}
\usage{
gkest1adjust(sigma = 1.5, data)
}
\arguments{
\item{sigma}{Smoothing Constant}

\item{filename}{Data file consisting of at least 3 columns: Maturity, Amount.issued, and Yield}
}
\value{
list consisting of a matrix of Target Tenor, sigma, Target Tenor Yield, Effective Term to Maturity, and Annual Yield; Interpolated 7 year yield (Semi-annual basis); Extrapolated 10 year yield (Semi-annual basis); Interpolated 7 year yield (Annual basis); and Extrapolated 10 year yield (Annual basis)
}
\description{
Interpolated and Extrapolated 7 and 10 year yields adjusting for Credit ratings
}

