% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gswt1.R
\name{gswt1}
\alias{gswt1}
\title{gswt1}
\usage{
gswt1(tt = 10, data = Jan15, sigma = 1.5, ...)
}
\arguments{
\item{tt}{Target Tenor}

\item{data}{Data file consisting of at least 3 columns: Maturity, Amount.issued, and Yield}

\item{sigma}{Smoothing Constant}

\item{maturity}{Maturity}

\item{issue}{Amount issued (should be numeric)}

\item{yield}{Yield}
}
\value{
a list of class "GaussianSmoother"
}
\description{
Local Constant Smoothing
}

