% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gswt1adjust.R
\name{gswt1adjust}
\alias{gswt1adjust}
\title{gswt1adjust}
\usage{
gswt1adjust(tt = 10, data = Jan15, sigma = 1.5)
}
\arguments{
\item{tt}{Target Tenor}

\item{data}{Data file consisting of at least 3 columns: Maturity, Amount.issued, and Yield}

\item{sigma}{Smoothing Constant}

\item{maturity}{Maturity}

\item{issue}{Amount issued}

\item{yield}{Yield}
}
\value{
a list of class "GaussianSmoothAdjust"
}
\description{
Local Constant Smoothing adjusted for Credit ratings
}

