\name{Jan15}
\alias{Jan15}
\docType{data}
\title{
Maturity, Yield, and Issue Amount for January 2015
}
\description{
Maturity, Yield, and Issue Amount for January 2015 for Bonds satisfying the RBA selection criteria
}
\usage{data("Jan15")}
\format{
  A data frame withat least 3 numeric columns.
  \describe{
    \item{\code{Maturity}}{a numeric vector}
    \item{\code{Yield}}{a numeric vector}
    \item{\code{Amount.issued}}{a numeric vector}
  }
}

\examples{
data(Jan15)
}
\keyword{datasets}
