% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lambdastart.R
\name{lambdastart}
\alias{lambdastart}
\title{Determining starting value for decay factor
Determining starting value for decay factor with Diebold and Li Parameterization}
\usage{
lambdastart(maturity = 2.5)
}
\arguments{
\item{maturity}{maturity at which second loading is maximized (in years).}
}
\value{
lambda maximizing second loading, to be used as starting value.
}
\description{
Determining starting value for decay factor
Determining starting value for decay factor with Diebold and Li Parameterization
}

