% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NelsonSiegel.R
\name{nelsonsiegelorig}
\alias{nelsonsiegelorig}
\title{Nelson-Siegel model
Nelson-Siegel model with Diebold and Li Parameterization}
\usage{
nelsonsiegelorig(maturity, lambda, beta1, beta2, beta3)
}
\arguments{
\item{maturity}{maturity in years}

\item{lambda}{decay factor}

\item{beta1}{long term factor}

\item{beta2}{short term factor}

\item{beta3}{curvature factor}
}
\value{
Nelson-Siegel function.
}
\description{
Nelson-Siegel model
Nelson-Siegel model with Diebold and Li Parameterization
}

