% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotGaussianSmooth.R
\name{plot.GaussianSmooth}
\alias{plot.GaussianSmooth}
\title{plot.GaussianSmooth}
\usage{
\method{plot}{GaussianSmooth}(obj, ...)
}
\arguments{
\item{obj}{an object of class "GaussianSmooth"}
}
\value{
A scatter plot of Yields vs Maturites with the Gaussian Smoothed line
}
\description{
plot method for class "GaussianSmooth"
}

