% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotGaussianSmoothadjust.R
\name{plot.GaussianSmoothAdjust}
\alias{plot.GaussianSmoothAdjust}
\title{plot.GaussianSmoothadjust}
\usage{
\method{plot}{GaussianSmoothAdjust}(obj)
}
\arguments{
\item{obj}{an object of class "GaussianSmoothAdjust"}
}
\value{
scatter plot of Yields vs Credit ratings with Gaussian smoother curves corresponding to the three credit rating BBB-, BBB, and BBB+
}
\description{
plot method for class "GaussianSmoothAdjust"
}

