% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotnelsonsiegel.R
\name{plot.NelsonSiegel}
\alias{plot.NelsonSiegel}
\title{plot.NelsonSiegel}
\usage{
\method{plot}{NelsonSiegel}(obj)
}
\arguments{
\item{obj}{an object of class "NelsonSiegel"}
}
\value{
scatter plot of Yields vs Maturities together with fitted Nelson-Siegel curve.
}
\description{
plot method for class "NelsonSiegel"
}

