% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotNelsonSiegeladjust.R
\name{plot.NelsonSiegelAdjust}
\alias{plot.NelsonSiegelAdjust}
\title{plot.NelsonSiegeladjust}
\usage{
\method{plot}{NelsonSiegelAdjust}(obj)
}
\arguments{
\item{obj}{an object of class "NelsonSiegelAdjust"}
}
\value{
scatter plot of Yields vs Maturities and fitted Nelson-Siegel curves corresponding to the three credit ratings BBB-, BBB, and BBB+
}
\description{
plot method for class "NelsonSiegelAdjust"
}

