% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotNelsonSiegelSvenssonadjust.R
\name{plot.NelsonSiegelSvenssonAdjust}
\alias{plot.NelsonSiegelSvenssonAdjust}
\title{plot.NelsonSiegelSvenssonadjust}
\usage{
\method{plot}{NelsonSiegelSvenssonAdjust}(obj)
}
\arguments{
\item{obj}{an object of class "NelsonSiegelSvenssonAdjust"}
}
\value{
scatter plot of Yields vs Maturities and fitted Nelson-Siegel-Svensson curves corresponding to the three credit ratings BBB-, BBB, and BBB+
}
\description{
plot method for class "NelsonSiegelSvenssonAdjust"
}

