% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/printGaussianSmooth.R
\name{print.GaussianSmooth}
\alias{print.GaussianSmooth}
\title{print.GaussianSmooth}
\usage{
\method{print}{GaussianSmooth}(obj, ...)
}
\arguments{
\item{obj}{an object of class "GaussianSmooth"}
}
\value{
numeric vector consisting of Target Tenor, sigma, Target Tenor Yield, and Effective Term to Maturity
}
\description{
print method for class "GaussianSmooth"
}

