% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/printGaussianSmoothadjust.R
\name{print.GaussianSmoothAdjust}
\alias{print.GaussianSmoothAdjust}
\title{print.GaussianSmoothadjust}
\usage{
\method{print}{GaussianSmoothAdjust}(obj)
}
\arguments{
\item{obj}{an object of class "GaussianSmoothAdjust"}
}
\value{
numeric vector consisting of Target Tenor, sigma, Target Tenor Yield, and Effective Term to Maturity
}
\description{
print method for class "GaussianSmoothAdjust"
}

