% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/printnelsonsiegel.R
\name{print.NelsonSiegel}
\alias{print.NelsonSiegel}
\title{print.NelsonSiegel}
\usage{
\method{print}{NelsonSiegel}(obj)
}
\arguments{
\item{obj}{an object of class "NelsonSiegel"}
}
\value{
list consisting of ten year yield (semi-annual basis) and ten year yield (annual basis)
}
\description{
print method for class "NelsonSiegel"
}

