% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/printNelsonSiegeladjust.R
\name{print.NelsonSiegelAdjust}
\alias{print.NelsonSiegelAdjust}
\title{print.NelsonSiegeladjust}
\usage{
\method{print}{NelsonSiegelAdjust}(obj)
}
\arguments{
\item{obj}{an object of class "NelsonSiegelAdjust"}
}
\value{
list consisting of  ten year yield (semi-annual basis)  and ten year yield (annual basis)
}
\description{
print method for class "NelsonSiegelAdjust"
}

