% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/printnelsonsiegelsvensson.R
\name{print.NelsonSiegelSvensson}
\alias{print.NelsonSiegelSvensson}
\title{print.NelsonSiegelSvensson}
\usage{
\method{print}{NelsonSiegelSvensson}(obj)
}
\arguments{
\item{obj}{an object of class "NelsonSiegelSvensson"}
}
\value{
list consisting of ten year yield (semi-annual basis) and ten year yield (annual basis)
}
\description{
print method for class "NelsonSiegelSvensson"
}

