% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/printNelsonSiegelSvenssonadjust.R
\name{print.NelsonSiegelSvenssonAdjust}
\alias{print.NelsonSiegelSvenssonAdjust}
\title{summary.NelsonSiegelSvenssonadjust}
\usage{
\method{print}{NelsonSiegelSvenssonAdjust}(obj)
}
\arguments{
\item{obj}{an object of class "NelsonSiegelSvenssonAdjust"}
}
\value{
list consisting of Fitted Nelson-Siegel-Svensson model, ten year yield (semi-annual basis),  and ten year yield (annual basis)
}
\description{
print method for class "NelsonSiegelSvenssonAdjust"
}

