% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summaryGaussianSmooth.R
\name{summary.GaussianSmooth}
\alias{summary.GaussianSmooth}
\title{summary.GaussianSmooth}
\usage{
\method{summary}{GaussianSmooth}(obj)
}
\arguments{
\item{obj}{an object of class "GaussianSmooth"}
}
\value{
list consisting of Annualized Yields at 3, 5, 7, 10 years term to maturity; Interpolated 7 year Yield (Semi-annual basis), Extrapolated 10 year Yield (Semi-annual basis), Interpolated 7 year Yield (Annualized Basis), and Extrapolated 10 year Yield (Annualized Basis))
}
\description{
summary method for class "GaussianSmooth"
}

