% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summaryGaussianSmoothadjust.R
\name{summary.GaussianSmoothAdjust}
\alias{summary.GaussianSmoothAdjust}
\title{summary.GaussianSmoothadjust}
\usage{
\method{summary}{GaussianSmoothAdjust}(obj)
}
\arguments{
\item{obj}{an object of class "GaussianSmoothAdjust"}
}
\value{
list consisting of Annualized Yields at 3, 5, 7, 10 years term to maturity; Interpolated 7 year Yield (Semi-annual basis), Extrapolated 10 year Yield (Semi-annual basis), Interpolated 7 year Yield (Annualized Basis), and Extrapolated 10 year Yield (Annualized Basis))
}
\description{
summary method for class "GaussianSmoothAdjust"
}

