% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summarynelsonsiegel.R
\name{summary.NelsonSiegel}
\alias{summary.NelsonSiegel}
\title{summary.NelsonSiegel}
\usage{
\method{summary}{NelsonSiegel}(obj)
}
\arguments{
\item{obj}{an object of class "NelsonSiegel"}
}
\value{
list consisting of fitted Nelson-Siegel model, ten year yield (semi-annual basis), and ten year yield (annual basis)
}
\description{
summary method for class "NelsonSiegel"
}

