% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summaryNelsonSiegeladjust.R
\name{summary.NelsonSiegelAdjust}
\alias{summary.NelsonSiegelAdjust}
\title{summary.NelsonSiegeladjust}
\usage{
\method{summary}{NelsonSiegelAdjust}(obj)
}
\arguments{
\item{obj}{an object of class "NelsonSiegelAdjust"}
}
\value{
list consisting of Fitted Nelson-Siegel model, ten year yield (semi-annual basis),  and ten year yield (annual basis)
}
\description{
summary method for class "NelsonSiegelAdjust"
}

