% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summarynelsonsiegelsvensson.R
\name{summary.NelsonSiegelSvensson}
\alias{summary.NelsonSiegelSvensson}
\title{summary.NelsonSiegelSvensson}
\usage{
\method{summary}{NelsonSiegelSvensson}(obj)
}
\arguments{
\item{obj}{an object of class "NelsonSiegel"}
}
\value{
list consisting of inverse decay factors, fitted Nelson-Siegel-Svensson model, ten year yield (semi-annual basis), and ten year yield (annual basis)
}
\description{
summary method for class "NelsonSiegelSvensson"
}

