% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summaryNelsonSiegelSvenssonadjust.R
\name{summary.NelsonSiegelSvenssonAdjust}
\alias{summary.NelsonSiegelSvenssonAdjust}
\title{summary.NelsonSiegelSvenssonadjust}
\usage{
\method{summary}{NelsonSiegelSvenssonAdjust}(obj)
}
\arguments{
\item{obj}{an object of class "NelsonSiegelSvenssonAdjust"}
}
\value{
list consisting of Fitted Nelson-Siegel-Svensson model, ten year yield (semi-annual basis),  and ten year yield (annual basis)
}
\description{
summary method for class "NelsonSiegelSvenssonAdjust"
}

